gfs2_edit: Avoid potential int overflow in find_journal_block()
Spotted by coverity: "Potentially overflowing expression ji.ji_nsegment
* 16U with type unsigned int (32 bits, unsigned) is evaluated using
32-bit arithmetic, and then used in a context that expects an expression
of type uint64_t"

+++ gfs2/edit/journal.c
@@ -72,7 +72,7 @@ uint64_t find_journal_block(const char *journal, uint64_t *j_size)
 		if (amtread) {
 			gfs_jindex_in(&ji, jbuf);
 			jblock = ji.ji_addr;
-			*j_size = ji.ji_nsegment * 0x10;
+			*j_size = (uint64_t)ji.ji_nsegment * 0x10;
 		}
 		inode_put(&jiinode);
 	} else {
